rename exception.py to exceptions.py
