+++ src/plugins/nonebot_plugin_pacman/exceptions.py
