+++ src/plugins/nonebot_plugin_wolfram_alpha/exceptions.py
